Rename package from runtime to compiler
